Refactor wildcard classes to use railway Responses
These classes are currently being used in the authentication
flow, but needed to be repurposed so they can be used in the
V2 Authenticators CRUD API.

In the V2 API, previously opaque error conditions related to
invalid wildcards need to be exposed to the user. The class
Response::Failure allows us to return a message from the
wildcard validation classes - the message will be ignored in
the authentication flow, but presented to the user in the
V2 API response.