Restrict wildcard matching in private domains
Private domains are those that are capable of issuing
subdomains to mutually un-trusting parties.

Our wildcard rules disallow wildcarding in DNS names
that would allow for matches across ownership boundaries.
For example, we restrict wildcards in eTLD+1 - all
subdomains with the same eTLD+1 are owned by the same
entity or organization.

This rule does not hold for private domains, which can
issue subdomains to mutually un-trusting parties. In this
context, a matches for a wildcarded subdomain cannot
guarantee shared ownership.